requirements.txt
# ==================================================
# Weapon Detection System - Requirements
# ==================================================

# Core Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.3.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Science & Analysis
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Dashboard & Visualization
streamlit>=1.28.0
plotly>=5.17.0
plotly-express>=0.4.1

# Database Integration
pymongo>=4.5.0
dnspython>=2.4.0

# HTTP & API
requests>=2.31.0

# Utilities
tqdm>=4.66.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# System Monitoring
psutil>=5.9.0

# Optional: Advanced Image Augmentation
# albumentations>=1.3.1

# Optional: Development & Testing
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.1.0
# mypy>=1.5.0

